InfoBubbleを表示する (#15)
* wip

* Fix: InfoWindow positioning is incorrect (confirmed: Google Maps, Mapbox, and Here. Not confirmed yet: ArcGIS)

* Info bubbles follow the dragged markers.

* update: .gitignore file

* wip: Implementing the marker drag gesture for Here SDK

* fix: linter errors

* fix: duplicated markers are added

* fix: KtLint errors

* update README.md

* The mapProvider of MapViewBase is no longer necessary

* Implement toasting logic in AppViewModel

* Create Settings class for defining default values

* rename MapView.kt to MapViewContainer.kt

* Fix: Cannot follow the map camera position for ArcGIS MapView

* Fix: InfoBubble does not follow the clicked marker on ArcGIS MapView

* Fix: onMapClick does not be involved

* Optimize code structure

* Fix: InfoBubble does not follow on ArcGIS Mapview when device is rotated

* Fix: linter errors

* Launch Google Maps app when taps on the "Get Directions" button in InfoBubble

+++ mapconductor-core/src/main/java/com/mapconductor/core/projection/WebMercator.kt
@@ -1,6 +1,6 @@
 package com.mapconductor.core.projection
 
-import com.mapconductor.core.Offset
+import androidx.compose.ui.geometry.Offset
 import com.mapconductor.core.features.IGeoPoint
 import kotlin.math.atan
 import kotlin.math.exp
@@ -11,7 +11,7 @@ object WebMercator : Projection {
     override fun project(position: IGeoPoint): Offset {
         val x = position.longitude * 20037508.34 / 180
         val y = ln(tan((90 + position.latitude) * Math.PI / 360)) * 20037508.34 / Math.PI
-        return Offset(x, y)
+        return Offset(x.toFloat(), y.toFloat())
     }
 
     override fun unproject(point: Offset): IGeoPoint {
